backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==2.0.12
diff-match-patch==20230430
Django==3.2
django-filter==23.5
djangorestframework==3.15.1
djangorestframework-simplejwt
djoser==2.1.0
drf-extra-fields==3.7.0
flake8==5.0.4
gunicorn==20.1.0
idna==3.7
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.0
pep8-naming==0.13.3
Pillow==9.3.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.9.1
pydocstyle==6.3.0
pyflakes==2.5.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2024.1
requests==2.26.0
snowballstemmer==2.2.0
sqlparse==0.5.0
tablib==3.5.0
toml==0.10.2
typing_extensions==4.11.0
urllib3==1.26.18
reportlab==4.2.0
hashids==1.3.1